--- pkg/common/const.go
@@ -1,23 +1,0 @@
-package common
-
-const SystemLogger = "system"
-const MetricPrefix = "i2eco"
-
-const ModAppName = "app"
-const ModLoggerName = "logger"
-const ModCmdName = "cmd"
-const ModPromName = "prom"
-const ModMysqlName = "mysql"
-const ModRedisName = "redis"
-const ModOssName = "oss"
-const ModMixCacheName = "mixcache"
-const ModMongoName = "mongo"
-const ModGinSessionName = "ginsession"
-const ModEchoSessionName = "echosession"
-const ModTplBeegoName = "tplbeego"
-const ModStatName = "stat"
-const ModGinName = "gin"
-const ModEtcdName = "etcd"
-const ModWechatName = "etcd"
-const ModTokenName = "token"
-const ModOpenGithubName = "opengithub"
